Infrastructure: Fix the TMxpSendTagHandlerTest and re-enable it (#4714)
* Fixes problem blocking compilation of TMxpSendTagHandlerTest

- Extracts TMediaData to a header in order to remove dependencies on MediaPlayer in testing environment

* Adds override specifier to TMxpStubClient functions
