+++ src/TMxpSoundTagHandler.cpp
@@ -18,7 +18,7 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 
-#include "TMedia.h"
+#include "TMediaData.h"
 #include "TMxpSoundTagHandler.h"
 #include "TMxpClient.h"
