+++ test/CMakeLists.txt
@@ -49,6 +49,5 @@ list(FILTER MXP_SOURCE EXCLUDE REGEX ".*/src/TMxpMudlet.cpp")
 add_executable(TMxpTagParserTest TMxpTagParserTest.cpp ../src/TMxpTagParser.cpp ../src/TMxpNodeBuilder.cpp ../src/MxpTag.cpp ../src/TStringUtils.cpp)
 add_test(NAME TMxpTagParserTest COMMAND TMxpTagParserTest)
 
-# https://github.com/Mudlet/Mudlet/issues/4021 added an include that made this test fail.  Entering issue to inspect the test separately; re-enable when complete.
-#add_executable(TMxpSendTagHandlerTest TMxpSendTagHandlerTest.cpp ${MXP_SOURCE} ../src/TMxpSendTagHandler.cpp ../src/TMxpTagParser.cpp ../src/TMxpNodeBuilder.cpp)
-#add_test(NAME TMxpSendTagHandlerTest COMMAND TMxpSendTagHandlerTest)
+add_executable(TMxpSendTagHandlerTest TMxpSendTagHandlerTest.cpp ${MXP_SOURCE} ../src/TMxpSendTagHandler.cpp ../src/TMxpTagParser.cpp ../src/TMxpNodeBuilder.cpp)
+add_test(NAME TMxpSendTagHandlerTest COMMAND TMxpSendTagHandlerTest)
