+++ test/TMxpStubClient.h
@@ -7,6 +7,7 @@
 
 #include "TMxpContext.h"
 #include "TMxpClient.h"
+#include "TMediaData.h"
 
 class TMxpStubContext : public TMxpContext {
 public:
@@ -23,12 +24,12 @@ public:
         return mSupportedElements;
     }
 
-    virtual TMxpTagHandlerResult handleTag(TMxpContext& ctx, TMxpClient& client, MxpTag* tag)
+    TMxpTagHandlerResult handleTag(TMxpContext& ctx, TMxpClient& client, MxpTag* tag) override
     {
         return MXP_TAG_HANDLED;
     }
 
-    virtual void handleContent(char ch)
+    void handleContent(char ch) override
     {
 
     }
